backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
groq>=0.4.2
PyPDF2>=3.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
